Created Module for plant related temperature measures: heatunits, chill units, days above threshold,  freeze hardening, temperature stress. Added reference to the temperatures for the simulation day from the climate input module so they no longer need to be passed. Changed the accumulated warm days crop climate parameter from integer to real since dividing in some place would result in trunction.
git-svn-id: https://infosys.ars.usda.gov/svn/code/weps1/branches/weps.src.subregion@14250 77af8a48-1994-4e4a-b6f4-b1c89642997b

+++ src/lib_structures/biomaterial.f95
@@ -56,7 +56,7 @@ module biomaterial
      real :: stemmasstrend  ! direction in which stem mass is trending.
                             ! Saves trend even if stem mass is static for long periods.
 
-     integer :: twarmdays      ! number of consecutive days that the temperature has been above the minimum growth temperature
+     real :: twarmdays      ! number of consecutive days that the temperature has been above the minimum growth temperature
      real :: tchillucum     ! accumulated chilling units (days)
      real :: thardnx        ! hardening index for winter annuals (range from 0 t0 2)
